Move IntRange out of vespajlib
vespajlib classes in the interface between config model and config server
causes LinkageErrors.
