+++ config-provisioning/src/main/java/com/yahoo/config/provision/Capacity.java
@@ -1,8 +1,6 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.config.provision;
 
-import com.yahoo.collections.IntRange;
-
 import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Stream;
