+++ config-provisioning/src/test/java/com/yahoo/config/provision/CapacityTest.java
@@ -1,7 +1,6 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.config.provision;
 
-import com.yahoo.collections.IntRange;
 import org.junit.jupiter.api.Test;
 
 import java.util.Optional;
