+++ config-provisioning/src/test/java/com/yahoo/config/provision/IntRangeTestCase.java
@@ -1,10 +1,10 @@
-package com.yahoo.collections;
+package com.yahoo.config.provision;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * @author bratseth
